[Struct]: spells up to 6.1.2

+++ src/server/game/Server/Packets/SpellPackets.h
@@ -348,7 +348,7 @@ namespace WorldPackets
 
             ObjectGuid CasterUnit;
             uint32 SpellID  = 0;
-            uint8 Reason    = 0;
+            uint16 Reason   = 0;
             uint8 CastID    = 0;
         };
 
@@ -361,7 +361,7 @@ namespace WorldPackets
 
             ObjectGuid CasterUnit;
             uint32 SpellID  = 0;
-            uint16 Reason   = 0;
+            uint8 Reason    = 0;
             uint8 CastID    = 0;
         };
